FastaLoad_Logic refactoring. Rund 200 Zeilen weniger Code dadurch. Zudem die Logic (loadFasta) umbenannt.
